backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
python-multipart==0.0.6
httpx==0.25.0
pytest==7.4.2
pytest-asyncio==0.21.1
requests==2.31.0
openai==0.28.0
langchain==0.0.292
tiktoken==0.5.1
pydub==0.25.1
jinja2==3.1.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 